buildflow-aws-image-classification/requirements.txt
buildflow
python-dotenv
cython
imageai
pandas
pillow>=7.0.0
numpy>=1.18.1
opencv-python>=4.1.2
ray
torch>=1.9.0 --extra-index-url https://download.pytorch.org/whl/cpu
torchvision>=0.10.0 --extra-index-url https://download.pytorch.org/whl/cpu
python-dotenv
pytest==7.1.3
tqdm==4.64.1
scipy>=1.7.3
matplotlib>=3.4.3
mock==4.0.3
s3fs
snowflake-connector-python
python-multipart
